Moving where the initialization step is
