+++ src/main/java/com/dougnoel/sentinel/steps/APISteps.java
@@ -192,4 +192,25 @@ public class APISteps {
 	public static void addHeader(String name, String value) {
 		APIManager.addHeader(name, SentinelStringUtils.replaceStoredVariables(value));
 	}
+
+	/**
+	 * Adds the parsed string keys and values in Configuration for later use
+	 *
+	 * @param values String the string with keys and values
+	 * Example:
+	 *  When I initialize the configuration values as follows
+	 *     """
+	 *     id: 10
+	 *     category_name: puppies
+	 *     """
+	 *
+	 */
+	@When("I initialize the configuration values as follows")
+	public void iInitializeTheData(String values) {
+		var items = values.replace(" ","").split("\n");
+		for (var item:items
+		) {
+			Configuration.update(item.split(":")[0],  item.split(":")[1]);
+		}
+	}
 }
